Shrink some enums so that trigger actions work
A bunch of these are negative values in the ini (high 24 bits all one). The effect of this was that win conditions didn't work and speaking accessed invalid indices

+++ CODE/taction.h
@@ -115,7 +115,7 @@ struct TActionClass {
 		QuarryType			Quarry;		// Preferred target for attack.
 		VQType				Movie;		// The movie to play.
 		bool					Bool;			// Boolean value.
-		long					Value;
+		int					Value;
 	} Data;
 
 	TActionClass(void) : Action(TACTION_NONE) {
